Edit/ ask_popup.dart 파일 최종 수정 noText 적지않으면 (null)로 두면 확인 버튼만 나오게 처리. 앞으로 팝업은 이거로 쓰면 될 것 같습니다
